Test class changed to Question class, Test class created, tests added to DataBase

+++ src/main/java/org/example/service/BaseService.java
@@ -2,7 +2,7 @@ package org.example.service;
 
 public interface BaseService<T,R> {
     T add(T t);
-    T get(int key);
+    T get(Long key);
     T upDate(T t, Long key);
-    R delete(int key);
+    R delete(Long key);
 }
