+++ src/main/java/org/example/service/TestService.java
@@ -6,24 +6,24 @@ import org.example.model.Test;
 public class TestService implements BaseService<Test,Boolean>{
     @Override
     public Test add(Test test) {
-       return DataBase.tests.put(test.getId(),test);
+        return DataBase.tests.put(test.getId(),test);
     }
 
     @Override
-    public Test get(int key) {
+    public Test get(Long key) {
         return DataBase.tests.get(key);
     }
 
     @Override
     public Test upDate(Test test, Long key) {
         if (DataBase.tests.get(key)==null){
-        return DataBase.tests.put(key,test);
+            return DataBase.tests.put(key, test);
         }
         return null;
     }
 
     @Override
-    public Boolean delete(int key) {
+    public Boolean delete(Long key) {
         DataBase.tests.get(key).setActive(false);
         return true;
     }
